feat: Add 'mise' workflow in GitHub actions
This commit introduces a new GitHub workflow for the 'mise' directory.
The workflow is triggered on push and pull request events that involve
changes in the 'mise' directory or the workflow file itself. It runs
on the latest version of Ubuntu and uses 'actions/checkout@v4' and
'jdx/mise-action@v2' actions.