feat(nvim-cmp): enhance key mappings for item selection and doc scrolling
The key mappings for item selection in nvim-cmp have been updated. Now,
"<C-j>" and "<C-k>" are used for selecting the next and previous items
respectively when the completion menu is visible. If the menu is not
visible, the default actions are performed.

Additionally, the key mappings for scrolling through documentation have
been changed. "<A-j>" and "<A-k>" are now used for scrolling down and
up through the documentation respectively.