feat(nvim): Update denippet and add autoimport
This commit updates the denippet plugin configuration and adds a new
plugin called denippet-autoimport-vscode. The denippet plugin
configuration is updated to include new events and dependencies. The
new plugin, denippet-autoimport-vscode, is added to the lazy-lock.json
file and its configuration is added to the denippet.lua file. Also,
a new snippet list is added to the ryoppippi-snippets-list.lua file.