feat(neovim): refactor neoconf and lspconfig for deno and vtsls
This commit refactors the neoconf and nvim-lspconfig configurations
specifically for deno and vtsls. It removes the dependency on
climbdir.nvim and introduces a new function to find the root directory
for deno. It also adds checks to stop vtsls if denols is working. The
commit also includes updates to utils.lua to add deno and node specific
files.