feat(nvim): Add vim-fall plugin
This commit adds the vim-fall plugin to the Neovim configuration. The
plugin is added to the lazy-lock.json file and a new configuration file
for the plugin is created in the lua/plugin directory. The plugin is
loaded lazily on the "User DenopsReady" event and depends on the
"vim-denops/denops.vim" plugin.