feat(nvim): add nvim-automa plugin
This commit adds the nvim-automa plugin to the Neovim configuration. The
plugin is added to the lazy-lock.json file with its corresponding branch
and commit hash. Additionally, a new Lua file is created for the plugin
under the nvim/lua/plugin directory. The plugin is set to load on the
"VeryLazy" event.