feat(denols): Refactor denols client stop logic
The logic to stop the denols client when vtsls or tsserver is running
has been refactored. The previous implementation was removed and a new
one was added in the init function. Now, the denols client is detached
from the buffer if vtsls or tsserver is running. If no buffer is
attached to denols, the client is stopped.