feat(nvim-lspconfig): add typescript-svelte-plugin to vtsls
This commit adds the typescript-svelte-plugin to the vtsls configuration
in Neovim's LSP config. The plugin is added to the globalPlugins section
of the tsserver configuration. This plugin enhances TypeScript support
for Svelte files.

The plugin is also added as a devDependency in the package.json file of
the node_servers directory. The corresponding lock file is updated as well.