feat(fish/config): Add auto venv activation
monaqaさんに教えてもらったよ
This commit introduces a new feature for automatic virtual environment
activation in fish shell. The script is based on a gist by bastibe, with
modifications to work with newer versions of fish and handle virtualenvs
not located in the root of a git directory. The function
`__auto_source_venv` is triggered on directory change and checks if the
current directory is inside a git directory. If so, it finds the closest
virtualenv starting from the current directory and activates it. If the
virtualenv is activated but we are not in a git directory, it deactivates
the virtualenv.