feat(venv.fish): Comment out auto venv sourcing function
The function `__auto_source_venv` in `venv.fish` has been commented out.
This function was responsible for automatically sourcing the virtual
environment upon directory change. The reason for this change is not
specified in the diff. If this function is no longer needed, consider
removing it entirely in a future commit.