feat: Update lackluster.nvim and set as default colorscheme
This commit updates the commit hash for lackluster.nvim in lazy-lock.json
indicating a version update. It also sets lackluster as the default
colorscheme in pack.lua. Additionally, it expands the configuration for
lackluster in lackluster.lua, adding a new autocommand for the ColorScheme
event and configuring various aspects of the colorscheme.