feat(kanagawa): improve kanagawa theme handling
This commit improves the handling of the kanagawa theme in Neovim. It
changes the isKanagawa function to check if the NVIM_COLORSCHEME starts
with "kanagawa" instead of being exactly "kanagawa". It also adds an
autocommand to set the colors_name variable when the ColorScheme event
is triggered with a scheme that starts with "kanagawa". The kanagawa
configuration is now done in the config function instead of a separate
kanagawaConfig function. The event that triggers the loading of the
kanagawa theme is changed from "UiEnter" to "ColorScheme".