feat(nvim): change quit command in CopilotChat
The quit command in the CopilotChat feature of the Neovim configuration
has been updated. Previously, it was set to 'wq' which only writes and
quits the current buffer. Now, it has been changed to 'wqa' which writes
all changed buffers and then quits Neovim. This change ensures that all
changes are saved before quitting.