feat(fish/config): Add launch editor scripts
This commit introduces two new scripts in the fish/config directory.
The first script, _launch_editor, is a utility script that opens a given
file at a specified line and column in the user's preferred editor. The
second script, launch_editor.fish, sets the LAUNCH_EDITOR environment
variable to the path of the _launch_editor script if it is not already
set. This allows the user to easily open files at specific locations
from the command line.