feat: Update plugin versions and integrate cmp with CopilotChat
This commit updates the versions of several plugins in the
lazy-lock.json file. Notably, the versions of CopilotChat.nvim,
copilot-cmp, copilot.lua, and gh-copilot have been updated.

Additionally, the CopilotChat plugin has been integrated with nvim-cmp
in the copilotChat.lua file. This integration is done in the init
function of the plugin configuration.