feat(nvim-cmp): enhance plugin loading and configuration
A new function `load_after` has been added to load plugins after they
are installed. This function is used in the configuration of each
dependency of `nvim-cmp`, ensuring that the plugins are loaded
correctly.

The `package.json` file in `node_servers` has also been updated with a
new `prepare` script.

The `bun.lockb` file has been updated with a new hash.

BREAKING CHANGE: The way plugins are loaded and configured has changed.
Existing configurations may need to be updated to use the new function.